Fix type mismatch when size_t is 32 bits.

+++ src/intcomp/CompressionFlags.h
@@ -48,8 +48,8 @@ struct AbbrevAssignFlags {
 };
 
 struct CompressionFlags {
-  uint64_t CountCutoff;
-  uint64_t WeightCutoff;
+  size_t CountCutoff;
+  size_t WeightCutoff;
   size_t LengthLimit;
   size_t MaxAbbreviations;
   decode::IntType SmallValueMax;
